requirements.txt
absl-py==1.3.0
beautifulsoup4==4.11.1
cachetools==5.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
docker-pycreds==0.4.0
filelock==3.9.0
fonttools==4.38.0
fvcore==0.1.5.post20221221
gdown==4.6.0
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.15.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
idna==3.4
imageio==2.24.0
importlib-metadata==6.0.0
iopath==0.1.10
joblib==1.2.0
kiwisolver==1.4.4
kornia==0.6.9
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
networkx==3.0
numpy==1.24.1
oauthlib==3.2.2
opencv-python==4.7.0.68
packaging==23.0
pathtools==0.1.2
Pillow==9.4.0
portalocker==2.6.0
promise==2.3
protobuf==3.20.3
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytorch3d==0.7.2
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.2.0
scipy==1.10.0
sentry-sdk==1.12.1
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
soft-renderer==1.0.0
soupsieve==2.3.2.post1
tabulate==0.9.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
threadpoolctl==3.1.0
tifffile==2022.10.10
torch==1.10.0+cu111
torchaudio==0.10.0+rocm4.1
torchvision==0.11.0+cu111
tqdm==4.64.1
trimesh==3.17.1
typing_extensions==4.4.0
urllib3==1.26.13
wandb==0.13.7
Werkzeug==2.2.2
yacs==0.1.8
zipp==3.11.0
